restructured profiles and other changes
