+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Home from "./pages/home/Home";
 import Login from "./pages/login/Login";
-import Profile from "./pages/profile-presets/Profile";
+import Profile from "./pages/profile/Profile";
 import Graph from "./pages/graph/Graph";
 import SignUpP from "./pages/signup/Signup";
 import "./App.css";
