+++ src/components/login/SignIn.jsx
@@ -34,7 +34,7 @@ export default function SignIn() {
         if(token) {
           console.log("Login successful");
           localStorage.setItem('token', token); 
-          window.location = '#/home';
+          window.location = '/#/home';
         } else {
           console.error("Invalid email or password");
         }
